Create a corresponding test file for each module
