+++ tests/test_util.py
@@ -2,13 +2,26 @@
 """
 Tests for gstorage.utils
 """
+import os
 import tempfile
+from mock import patch
 from unittest import TestCase
 
-from gstorage.utils import find_files
+from django.conf import settings
+
+from gstorage.apps import GStorageConfig
+from gstorage.utils import find_files, get_config
+
+key = 'GOOGLE_APPLICATION_CREDENTIALS'
 
 
 class TestUtil(TestCase):
+
+    def setUp(self):
+        try:
+            del os.environ[key]
+        except KeyError:
+            pass
 
     def test_find_files_empty_directory(self):
         path = tempfile.mkdtemp()
@@ -26,3 +39,18 @@ class TestUtil(TestCase):
         _, child_filename = tempfile.mkstemp(dir=child)
 
         assert find_files(path) == [filename, child_filename]
+
+    @patch('gstorage.checks.REQUIRED_SETTINGS', [key])
+    def test_key_override(self):
+        """Test that config retains the value in settings over environment"""
+        setattr(settings, key, '/foo')
+        os.environ[key] = '/bar'
+        assert get_config(key) == '/foo'
+
+    @patch('gstorage.checks.REQUIRED_SETTINGS', [key])
+    def test_real_config(self):
+        """Test that instantiating the real config calls the validation methods"""
+        with patch('gstorage.apps.checks.register') as mock_method:
+            app = GStorageConfig.create('gstorage')
+            app.ready()
+            assert mock_method.call_count == 1
